Chapter 10 - Best Practices and Coding Style/Mega Projects/Multilingual AI Assistant/requirements.txt
SpeechRecognition
pyaudio
google-generativeai
gTTS
pipwin
streamlit